% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_c14_matrix.R
\name{build_c14_matrix}
\alias{build_c14_matrix}
\title{Build C14 Matrix.}
\usage{
build_c14_matrix(c14_dates, BP = TRUE, resolution = -1)
}
\arguments{
\item{c14_dates}{A matrix with two columns. The first column should contain
mean uncalibrated radiocarbon dates and the second should contain the
associated errors.}

\item{BP}{Logical. Use Before Present (BP) timescale?}

\item{resolution}{Desired temporal resolution in years.}
}
\value{
A list with two elements. The first element is a chronological
uncertainty matrix containing a number of columns equal to the number of
dates (rows) in the c14_matrix. Each column contains a single calibrated
radiocarbon date density. The rows contain samples of the corresponding
calibrated date density. The second element in the returned list contains
the time range spanned by all calibrated dates. This range and the
resolution argument can be used to recover time-stamps for each column in
the chronological uncertainty matrix.
}
\description{
Build C14 Matrix.
}
