% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_events.R
\name{count_events}
\alias{count_events}
\title{Count events.}
\usage{
count_events(x, breaks, BP)
}
\arguments{
\item{x}{A numeric vector of event times to be counted.}

\item{breaks}{A numeric vector of temporal bin edges The bin edges defined
by this argument will serve as the right-most boundary condition which will
be closed, i.e., the interval will be left-open and right-closed: ( ].}

\item{BP}{A logical argument specifying whether the events are on the BP
(before present) timeline.}
}
\value{
A vector containing event counts.
}
\description{
Counts the number of events from 'x' that fall into time-bins defined by
breaks.
}
