% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/public_utils.R
\name{find_max_count}
\alias{find_max_count}
\title{Find max count.}
\usage{
find_max_count(event_count_freqs)
}
\arguments{
\item{event_count_freqs}{A matrix where each column refers to the count of
events and the rows refer to times (temporal bins used for counting). The
cells each indicate the number of times a given count--time pair has been
sampled.}
}
\value{
Integer indicating the maximum count sampled from any temporal bin in
the event_count_freqs matrix.
}
\description{
Find max count.
}
