% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_count_ensemble.R
\name{plot_count_ensemble}
\alias{plot_count_ensemble}
\title{Plot count ensemble.}
\usage{
plot_count_ensemble(
  count_ensemble,
  times,
  use_ggplot2 = FALSE,
  axis_x_res = 100,
  axis_y_res = 1
)
}
\arguments{
\item{count_ensemble}{A matrix with comlumns that each refer to a probable
count sequence and rows refer to the count of events at a given time.}

\item{times}{A vector of times stamps that correspond to the times of the
count_ensemble rows.}

\item{use_ggplot2}{Logical. Should the package ggplot2 be used for plotting?}

\item{axis_x_res}{The resolution for the plotted x axis (time). Not used if
use_ggplot2 == T.}

\item{axis_y_res}{The resolution for the plotted y axis (count). Not used if
use_ggplot2 == T.}
}
\value{
This plot function returns NULL if use_ggplot2 == FALSE or a ggplot2
object if use_ggplot2 == T.
}
\description{
Plot count ensemble.
}
