% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rece.R
\name{rece}
\alias{rece}
\title{Build and optionally plot a radiocarbon-dated event count ensemble (RECE).}
\usage{
rece(
  c14_dates,
  nsamples,
  breaks,
  verbose = TRUE,
  BP = TRUE,
  cal_resolution = -1,
  plot_it = TRUE,
  time_axis_index = 100
)
}
\arguments{
\item{c14_dates}{A matrix with two columns. The first column should contain
mean uncalibrated radiocarbon dates and the second should contain the
associated errors.}

\item{nsamples}{Integer number of sampled sequences to plot.}

\item{breaks}{Vector of temporal breakpoints to use for sampling. This is
the analytical/plot resolution, not the calibration resolution. It also
defines the temporal boundaries of the RECE.}

\item{verbose}{Logical. The plot takes longer to create for more
dates/samples. So, this logical parameter indicates whether the user wants
to see messages indicating progress.}

\item{BP}{Logical. Use Before Present (BP) timescale?}

\item{cal_resolution}{Desired temporal resolution for calibration in years.}

\item{time_axis_index}{Integer value determining which elements of
breaks vector will be included in the x-axis ticks and labels for the
optional plot. The axis ticks will refer to time bin midpoints and be drawn
and labelled at seq(1, length(breaks) - 1, time_axis_label_index).}
}
\value{
Invisibly returns a list with three elements. The first is an
radiocarbon-dated event count ensemble, which is a matrix containing
probable count sequences where each column contains one probable sequence.
The second element is matrix containing frequencies of counts in each time
bin defined by breaks (used in the RECE heat map plot). The last element is
the breaks argument for convenience---it can be used to provide timestamps
for the first two returned objects (corresponds to the rows in each case).
}
\description{
Build and optionally plot a radiocarbon-dated event count ensemble (RECE).
}
