% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateEventCounts.R
\name{simulateEventCounts}
\alias{simulateEventCounts}
\title{Simulate event counts.}
\usage{
simulateEventCounts(
  process,
  times,
  nevents,
  nsamples,
  binning_resolution = -1,
  BP = T,
  c14 = T,
  ceMatrix = NULL,
  bigmatrix = T,
  parallel = T
)
}
\arguments{
\item{process}{A vector describing the background event-count process (i.e.,
conditional mean of count-based random variable). This will be passed to
R's 'sample()' function as the 'prob' parameter.}

\item{times}{A vector of event times to sample from.}

\item{nevents}{Integer number of events with uncertain event-times.}

\item{nsamples}{Integer number of sample sequences to draw.}

\item{binning_resolution}{Resolution of the desired time intervals (time
bins). It should be negative if using the BP timescale (i.e, if BP = T).}

\item{BP}{Logical (default T). Assume a Before Present timescale?}

\item{c14}{Logical (default T). Are the events dated with radiocarbon? If
so, the R package 'clam' will be used to simulate c14 dates from the 'true'
samples of the 'times' vector and then calibrate those dates.}

\item{ceMatrix}{A matrix containing discrete estimates describing
chronological uncertainty. Each column should contain density estimates
for a single event and the rows should each refer to discrete times. If c14
= F, then this matrix must be included.}

\item{bigmatrix}{A character vector containing a path pointing to a
'bigmemory' matrix descriptor file, or NULL (default).}

\item{parallel}{Logical (default = T). Use parallel (multiple processors) to
speed up computation?}
}
\value{
A list with 1) a matrix containing probable event count sequences
'(or a file path pointing to a file-backed bigmatrix); 2) a dataframe
containing the true (chronological-error-free) event-count sample.
}
\description{
Simulate event counts.
}
