% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_frequencies.R
\name{tabulate_freqs}
\alias{tabulate_freqs}
\title{Tabulate frequency distribution of event counts for a given interval.}
\usage{
tabulate_freqs(event_count_sample, nevents)
}
\arguments{
\item{event_count_sample}{An integer vector containing samples of event
count sequences for a given interval. Each element should refer to the
number of times events fell into the relevant interval.}

\item{nevents}{Number of events.}
}
\value{
A vector containing frequencies of event counts for each interval.
Each element will refer to the number of times a given count was
sampled for a given time.
}
\description{
Tabulate frequency distribution of event counts for a given interval.
}
